validation/Fig5/Fig5_LargeST/requirements.txt
easy-torch==1.3.2
easydict==1.10
pandas==1.3.5
packaging==23.1
setuptools==59.5.0
scipy==1.7.3
tables==3.7.0
sympy==1.10.1
setproctitle==1.3.2
scikit-learn==1.0.2
einops==0.6.1
matplotlib==3.8.2
numpy==1.22.4
networkx==2.6.3